Items - validation

+++ resources/views/welcome.blade.php
@@ -1,29 +1,12 @@
-<!doctype html>
-<html lang="{{ app()->getLocale() }}">
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <title>Cart</title>
-
-        <link href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.2/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-PsH8R72JQ3SOdhVi3uxftmaW6Vc51MKb0q5P2rRUpPvrszuE4W1povHYgTpBfshb" crossorigin="anonymous">
-        <style media="screen">
-            #page {
-                margin: 40px;
-            }
-            ul  {
-                list-style: none;
-            }
-            li {
-                padding: 20px;
-                border: 1px solid #bdbdbd;
-            }
-        </style>
-    </head>
-    <body>
-        <div class="container" id="page">
-            <div class="row">
+@extends('layout')
+@section('title')
+    Cart
+@endsection
+@section('content')
+    <a href="{{route('items.index')}}" class="btn btn-primary">
+        Manage Items
+    </a>
+    <div class="row">
                 <div class="col-md-6" id="items">
                     <h3 class="text-center">Items</h3>
 
@@ -31,19 +14,33 @@
                         @if(count($items))
                             @foreach($items as $item)
                                 <li data-id={{$item->getId()->toString()}}>
-                                    <span data-type="name{{$item->getId()->toString()}}">
-                                        {{$item->getName()}}
+                                    <span class="row">
+                                        <span class="col-md-6">
+                                            <span data-type="name{{$item->getId()->toString()}}">
+                                                {{$item->getName()}}
+                                            </span>
+                                            <span data-type="price{{$item->getId()->toString()}}">
+                                                ${{$item->getPrice()}}
+                                            </span>
+                                            @if($item->getSale())
+                                            <span class="text-danger">*on sale</span>
+                                            @endif
+                                        </span>
+
+                                        <span class="col-md-6">
+                                            <a href="#" data-type="cart" class="btn btn-sm btn-primary">cart</a>
+                                            <a href="#" data-type="wishlist" class="btn btn-sm btn-info">wishlist</a>
+                                        </span>
                                     </span>
-                                    <span data-type="price{{$item->getId()->toString()}}">
-                                        ${{$item->getPrice()}}
-                                    </span>
-                                    @if($item->getSale())
-                                        <span class="text-danger">*on sale</span>
-                                    @endif
-                                    <a href="#" data-type="cart" class="btn btn-sm btn-primary">cart</a>
-                                    <a href="#" data-type="wishlist" class="btn btn-sm btn-info">wishlist</a>
                                 </li>
                             @endforeach
+                        @else
+                            <li>
+                                No items,
+                                <a href="{{route('items.create')}}" class="btn btn-primary">
+                                    create a new item
+                                </a>
+                            </li>
                         @endif
                     </ul>
                 </div>
@@ -84,32 +81,44 @@
                     </div>
                 </div>
             </div>
-        </div>
+@endsection
 
-        <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
-        <script type="text/javascript" src="{{asset('js/functions.js')}}"></script>
-        <script type="text/javascript">
+@section('scripts')
+    <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossorigin="anonymous"></script>
+    <script type="text/javascript" src="{{asset('js/functions.js')}}"></script>
+    <script type="text/javascript">
             $("#items li a").on('click', function(){
                 var item_id = $(this).closest('li').data('id');
                 var item_name = $(`span[data-type="name${item_id}"]`).html();
                 var item_price = $(`span[data-type="price${item_id}"]`).html();
                 var item_type = $(this).data('type');
                 var cart_id = $(`#${item_type}`).data('id');
-                console.log(cart_id);
 
-                if ($(`#${item_type} ul li`).hasClass('empty')) {
-                    $(`#${item_type} ul`).html('');
+                if ($(`#${item_type} ul li[data-id="${item_id}"]`).length == 0) {
+
+                    if ($(`#${item_type} ul li`).hasClass('empty')) {
+                        $(`#${item_type} ul`).html('');
+                    }
+
+                    $(`#${item_type} ul`).append(`<li data-id=${item_id}>
+                        <h5>
+                        <span class="row">
+                        <span class="col-md-6">
+                        <span>${item_name}</span>
+                        <small>${item_price}</small>
+                        </span>
+                        <span class="col-md-6">
+                        <a href="#" data-type="delete" class="btn btn-sm btn-danger">remove</a>
+                        </span>
+                        </span>
+                        </h5>
+                        </li>`);
+
+                        addItem(item_id, cart_id);
+                } else {
+                    alert("Can't add item twice");
                 }
 
-                $(`#${item_type} ul`).append(`<li data-id=${item_id}>
-                                                <h5>
-                                                    <span>${item_name}</span>
-                                                    <small>${item_price}</small>
-                                                    <a href="#" data-type="delete" class="btn btn-sm btn-danger">remove</a>
-                                                </h5>
-                                            </li>`);
-
-                addItem(item_id, cart_id);
             });
 
             $("ul").on('click', "a[data-type='delete']", function () {
@@ -145,5 +154,4 @@
 
 
         </script>
-    </body>
-</html>
+@endsection
